+++ app/src/docs/IconContainer.doc.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import * as uui from '@epam/uui';
 import * as loveship from '@epam/loveship';
 import * as promo from '@epam/promo';
+import * as electric from '@epam/electric';
 import { TDocConfig, TDocContext, TSkin } from '@epam/uui-docs';
 import { EditableDocContent, DocExample, BaseDocsBlock } from '../common';
 
@@ -13,6 +14,7 @@ export class IconContainerDoc extends BaseDocsBlock {
         contexts: [TDocContext.Default, TDocContext.Form],
         bySkin: {
             [TSkin.UUI]: { type: '@epam/uui-components:ControlIconProps', component: uui.IconContainer },
+            [TSkin.Electric]: { type: '@epam/uui-components:ControlIconProps', component: electric.IconContainer },
             [TSkin.Loveship]: { type: '@epam/loveship:IconContainerProps', component: loveship.IconContainer },
             [TSkin.Promo]: { type: '@epam/promo:IconContainerProps', component: promo.IconContainer },
         },