+++ app/src/docs/LabeledInput.doc.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import * as uui from '@epam/uui';
 import * as promo from '@epam/promo';
 import * as loveship from '@epam/loveship';
+import * as electric from '@epam/electric';
 import { DocBuilder, TDocConfig, TDocContext, TSkin } from '@epam/uui-docs';
 import { BaseDocsBlock, DocExample, EditableDocContent } from '../common';
 
@@ -13,6 +14,7 @@ export class LabeledInputDoc extends BaseDocsBlock {
         contexts: [TDocContext.Default, TDocContext.Resizable, TDocContext.Form],
         bySkin: {
             [TSkin.UUI]: { type: '@epam/uui:LabeledInputProps', component: uui.LabeledInput },
+            [TSkin.Electric]: { type: '@epam/uui:LabeledInputProps', component: electric.LabeledInput },
             [TSkin.Loveship]: { type: '@epam/uui:LabeledInputProps', component: loveship.LabeledInput },
             [TSkin.Promo]: { type: '@epam/uui:LabeledInputProps', component: promo.LabeledInput },
         },